Test setup * Test files moved to test/ * test copy the needed files * some tests starts in own context (remote-service, messaging)
