+++ package.json
@@ -18,33 +18,32 @@
   "devDependencies": {
     "@cap-js/audit-logging": "^0.8.0",
     "@cap-js/change-tracking": "^1.0.6",
-    "@sap-cloud-sdk/connectivity": "^3.16.0",
-    "@sap-cloud-sdk/http-client": "^3.16.0",
-    "@sap-cloud-sdk/resilience": "^3.16.0",
-    "@sap-cloud-sdk/util": "^3.16.0",
     "@cap-js/sqlite": "^1.0.1",
     "@sap/ux-specification": "^1.108.4",
     "axios": "^1.4.0",
     "chai": "^4.3.7",
     "chai-as-promised": "^7.1.1",
-    "chai-subset": "^1.6.0"
+    "chai-subset": "^1.6.0",
+    "shx": "^0.3.4"
   },
   "scripts": {
     "watch": "cds watch",
     "start": "cds-serve",
     "test": "npx jest --silent",
-    "add-change-tracking": "npm add @cap-js/change-tracking && cp xmpls/change-tracking.cds ./srv && cp xmpls/change-tracking.test.js ./test",
+    "add-change-tracking": "npm add @cap-js/change-tracking && cp xmpls/change-tracking.cds ./srv",
     "add-telemetry": "npm add @cap-js/telemetry",
     "add-attachments": "npm add @cap-js/attachments && cp -r xmpls/attachments.cds ./db",
     "add-notifications": "npm add @cap-js/notifications && cp xmpls/alert-notifications.js ./srv && cp xmpls/notification-types.json ./srv",
-    "add-audit-log": "npm add @cap-js/audit-logging && cp xmpls/data-privacy.cds ./srv && cp xmpls/audit-log.test.js ./test",
+    "add-audit-log": "npm add @cap-js/audit-logging && cp xmpls/data-privacy.cds ./srv",
     "add-remote-service": "./.github/workflows/checkout remote-service",
     "add-messaging": "./.github/workflows/checkout messaging",
     "add-all-xmpls": "npm run add-remote-service && npm run add-messaging && npm run add-change-tracking && npm run add-audit-log",
     "reset": "read -p 'This will irreversibly reset your working directory including ALL files in this git repo. Continue?' -n 1 -r && echo && if [[ $REPLY =~ ^[Yy]$ ]]; then git clean -fd && git reset --hard && npm i; fi"
   },
   "jest": {
-    "modulePathIgnorePatterns": ["<rootDir>/xmpls/"]
+    "modulePathIgnorePatterns": [
+      "<rootDir>/xmpls/"
+    ]
   },
   "sapux": [
     "app/incidents"
