+++ test/messaging.test.js
@@ -1,10 +1,12 @@
 const cds = require("@sap/cds");
-const { GET, POST, PATCH, DELETE, expect, axios } = cds.test(__dirname + '../../', '--with-mocks');
-axios.defaults.auth = { username: "alice" };
+describe("Integration Test for Eventing", () => {
 
-describe("Integration Test for Remote Service", () => {
-    let draftId,incidentId;
-    describe("Test the BusinessPartner GET Endpoints", () => {
+  const { GET, POST, PATCH, expect, axios } = cds.test(__dirname + '/../xmpls/messaging', '--with-mocks');
+
+  axios.defaults.auth = { username: "alice" };
+  let draftId,incidentId;
+
+    describe("GET should return 200", () => {
         
         it("Should return list of Business Partners", async () => {
         const response = await GET("/odata/v4/api-business-partner/A_BusinessPartner");
@@ -45,7 +47,6 @@ describe("Integration Test for Remote Service", () => {
       });
       it('+ Test the customer detail', async () => {
         const response = await GET(`/odata/v4/processor/Incidents?$filter=ID eq ${draftId}`);
-        //incidentId = ID;
         expect(response.status).to.eql(200);
         expect(response.data.value).to.exist;
         expect(response.data.value[0]).to.contains({
@@ -55,24 +56,44 @@ describe("Integration Test for Remote Service", () => {
       });
    
         describe("Create annd Update Business Partner", () => {
-            it("Creates a new Business Partner", async () => {
-            const payload = {
-                BusinessPartner: "17100015",
-                BusinessPartnerIsBlocked: true,
-                BusinessPartnerFullName: "John Doee",
-            };
-            const response = await POST(
-                "/odata/v4/api-business-partner/A_BusinessPartner",
-                payload
-            );
-            expect(response.status).to.eql(201);
-            });
             it("Update Business Partner", async () => {
                 const response = await PATCH(
-                `/odata/v4/api-business-partner/A_BusinessPartner('17100015')`,
-                {BusinessPartnerIsBlocked: false}
+                `/odata/v4/api-business-partner/A_BusinessPartner('1004100')`,
+                {
+                  to_BusinessPartnerAddress: [{
+                     AddressID: "457",
+                     to_EmailAddress:[{
+                         AddressID: "457",
+                         Person: "johnson",
+                         OrdinalNumber: "334",
+                         EmailAddress: "sunny@test.com"
+                     }]
+                 }]
+             }
                 );
                 expect(response.status).to.eql(200);
+            });
+            describe("Verify the updated Business Partner", () => {
+              it("Verify the Address of Business Partner", async () => {
+                const response = await GET(`/odata/v4/api-business-partner/A_BusinessPartnerAddress?$filter=BusinessPartner eq '1004100'`);
+                expect(response.status).to.eql(200);
+                expect(response.data.value).to.exist;
+                expect(response.data.value[0]).to.contains({
+                  AddressID: "457",
+                });
+              }); 
+
+              it("Verify the Email address of Business Partner", async () => {
+                const response = await GET(`/odata/v4/api-business-partner/A_AddressEmailAddress?$filter=AddressID eq '457'`);
+                expect(response.status).to.eql(200);
+                expect(response.data.value).to.exist;
+                expect(response.data.value[0]).to.contains({
+                  AddressID: "457",
+                  Person: "johnson",
+                  OrdinalNumber: "334",
+                  EmailAddress: "sunny@test.com"
+                });
+              }); 
             });
         }); 
          
@@ -82,10 +103,6 @@ describe("Integration Test for Remote Service", () => {
                 "PreserveChanges": true
                 });
                 expect(status).to.equal(201);
-            }); 
-            it(`Update Business Partner details of the Incident`, async ()=>{
-                const {status } = await PATCH(`/odata/v4/processor/Incidents(ID=${draftId},IsActiveEntity=false)`,{customer_ID: '17100015'});
-                expect(status).to.equal(200);
             }); 
     });
 });
