--- xmpls/messaging/app/fields.cds
@@ -1,1 +1,0 @@
-using from './incidents/field';
