+++ xmpls/remote-service/srv/mashup.cds
@@ -1,0 +1,4 @@
+using { ProcessorService } from '../../../srv/services';
+using from '../../../srv/services';
+using from '../../../app/services';
+annotate ProcessorService with @impl: '../xmpls/remote-service/srv/services.js';
